+++ ui/src/features/keys/select-theme/lib/constants.ts
@@ -3,4 +3,5 @@ import type { KeyboardThemeOption } from '../model'
 export const themeOptions: KeyboardThemeOption[] = [
   { title: 'Default', value: 'defaultTheme' },
   { title: 'Light Gray And White', value: 'lightGrayAndWhiteTheme' },
+  { title: 'Matcha', value: 'matchaTheme' },
 ]